split useRecentVenueUsers into it's own file
